feature: flags for language selection
Closes https://github.com/felix-d1strict/LogiReader/issues/3
